Migrate GaloisRingSharedIris to protocol submodule
