+++ iris-mpc-cpu/benches/hnsw.rs
@@ -2,13 +2,11 @@ use aes_prng::AesRng;
 use criterion::{black_box, criterion_group, criterion_main, BenchmarkId, Criterion, SamplingMode};
 use iris_mpc_common::iris_db::{db::IrisDB, iris::IrisCode};
 use iris_mpc_cpu::{
-    database_generators::{create_random_sharing, generate_galois_iris_shares},
-    execution::local::LocalRuntime,
-    hawkers::{aby3_store::Aby3Store, plaintext_store::PlaintextStore},
-    hnsw::{GraphMem, HnswSearcher},
-    protocol::ops::{
-        batch_signed_lift_vec, cross_compare, galois_ring_pairwise_distance, galois_ring_to_rep3,
-    },
+    database_generators::create_random_sharing, execution::local::LocalRuntime, hawkers::{aby3_store::Aby3Store, plaintext_store::PlaintextStore}, hnsw::{GraphMem, HnswSearcher}, protocol::{
+        gr_shared_iris::GaloisRingSharedIris, ops::{
+            batch_signed_lift_vec, cross_compare, galois_ring_pairwise_distance, galois_ring_to_rep3,
+        }
+    }
 };
 use rand::SeedableRng;
 use tokio::task::JoinSet;
@@ -117,10 +115,10 @@ fn bench_gr_primitives(c: &mut Criterion) {
             let mut rng = AesRng::seed_from_u64(0);
             let iris_db = IrisDB::new_random_rng(4, &mut rng).db;
 
-            let x1 = generate_galois_iris_shares(&mut rng, iris_db[0].clone());
-            let y1 = generate_galois_iris_shares(&mut rng, iris_db[2].clone());
-            let x2 = generate_galois_iris_shares(&mut rng, iris_db[1].clone());
-            let y2 = generate_galois_iris_shares(&mut rng, iris_db[3].clone());
+            let x1 = GaloisRingSharedIris::generate_galois_iris_shares(&mut rng, iris_db[0].clone());
+            let y1 = GaloisRingSharedIris::generate_galois_iris_shares(&mut rng, iris_db[2].clone());
+            let x2 = GaloisRingSharedIris::generate_galois_iris_shares(&mut rng, iris_db[1].clone());
+            let y2 = GaloisRingSharedIris::generate_galois_iris_shares(&mut rng, iris_db[3].clone());
 
             let mut jobs = JoinSet::new();
             for (index, player) in runtime.get_identities().iter().enumerate() {
@@ -204,7 +202,7 @@ fn bench_gr_ready_made_hnsw(c: &mut Criterion) {
                         let searcher = HnswSearcher::default();
                         let mut rng = AesRng::seed_from_u64(0_u64);
                         let on_the_fly_query = IrisDB::new_random_rng(1, &mut rng).db[0].clone();
-                        let raw_query = generate_galois_iris_shares(&mut rng, on_the_fly_query);
+                        let raw_query = GaloisRingSharedIris::generate_galois_iris_shares(&mut rng, on_the_fly_query);
 
                         let mut jobs = JoinSet::new();
 
@@ -238,7 +236,7 @@ fn bench_gr_ready_made_hnsw(c: &mut Criterion) {
                         let searcher = HnswSearcher::default();
                         let mut rng = AesRng::seed_from_u64(0_u64);
                         let on_the_fly_query = IrisDB::new_random_rng(1, &mut rng).db[0].clone();
-                        let raw_query = generate_galois_iris_shares(&mut rng, on_the_fly_query);
+                        let raw_query = GaloisRingSharedIris::generate_galois_iris_shares(&mut rng, on_the_fly_query);
 
                         let mut jobs = JoinSet::new();
                         for (vector_store, graph_store) in vectors_graphs.into_iter() {
