+++ iris-mpc-cpu/src/execution/hawk_main.rs
@@ -1,6 +1,5 @@
 use super::player::Identity;
 use crate::{
-    database_generators::GaloisRingSharedIris,
     execution::{
         local::generate_local_identities,
         player::{Role, RoleAssignment},
@@ -14,7 +13,10 @@ use crate::{
     },
     network::grpc::{GrpcConfig, GrpcNetworking},
     proto_generated::party_node::party_node_server::PartyNodeServer,
-    protocol::ops::setup_replicated_prf,
+    protocol::{
+        ops::setup_replicated_prf,
+        gr_shared_iris::GaloisRingSharedIris,
+    }
 };
 use aes_prng::AesRng;
 use clap::Parser;
@@ -744,8 +746,8 @@ pub async fn hawk_main(args: HawkArgs) -> Result<HawkHandle> {
 mod tests {
     use super::*;
     use crate::{
-        database_generators::generate_galois_iris_shares,
         execution::local::get_free_local_addresses,
+        protocol::gr_shared_iris::GaloisRingSharedIris,
     };
     use iris_mpc_common::{
         helpers::smpc_request::UNIQUENESS_MESSAGE_TYPE, iris_db::db::IrisDB, job::BatchMetadata,
@@ -792,7 +794,7 @@ mod tests {
         let irises = IrisDB::new_random_rng(batch_size, iris_rng)
             .db
             .into_iter()
-            .map(|iris| generate_galois_iris_shares(iris_rng, iris))
+            .map(|iris| GaloisRingSharedIris::generate_galois_iris_shares(iris_rng, iris))
             .collect_vec();
         // Unzip: party -> iris_id -> share
         let irises = (0..n_parties)
