+++ iris-mpc-cpu/src/protocol/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod binary;
 pub mod ops;
 pub(crate) mod prf;
+pub mod gr_shared_iris;
